pyproject.toml
[project]
name = "dreamydormouse"
version = "0.1.0"
description = "RAG system for querying Markdown files"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    # Use lightweight alternatives
    "chromadb>=0.4.0",  # Vector database
    "tiktoken>=0.5.0",   # Token counting
    "pydantic>=2.0.0",   # Data validation
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

